[LayoutNG] Add percentage_resolution_block_size into the min/max pass.
This was surprising to me! But the min/max sizes depend on the
percentage resolution block-size, as replaced elements reflect this
size into their width.

This was working previously as the replaced element was manually
computing the heights, but this adds it into the pass explicitly.

This fixes it for absolute positioned objects which use the current
final height as the input to this function.

Eventually we'll want to cache based off this, and record if there is
a replaced element with percentage block-size in the calculation chain.

Bug: 635619
Change-Id: I6db9a90007b6c7d70f51d5ae1209ffba8a629e03